// src/leads/enums/lead.enums.ts
export enum LeadSource {
  RECOMENDACION = 'Recomendación',
  REDES_SOCIALES = 'Redes sociales',
  CONSULTA_WEB = 'Consulta web',
}

export enum LeadStatus {
  NUEVO = 'Nuevo',
  CONTACTADO = 'Contactado',
  EN_SEGUIMIENTO = 'En Seguimiento',
  NEGOCIACION = 'Negociación',
  CERRADO = 'Cerrado',
  PERDIDO = 'Perdido',
  NO_APTO = 'No Apto',
}